replace java.lang.reflect.AccessibleObject.setAccessible(boolean) by ReflectUtils.makeAccess
