+++ dubbo-common/src/main/java/org/apache/dubbo/common/utils/ReflectUtils.java
@@ -1381,4 +1381,17 @@ public final class ReflectUtils {
             ctor.setAccessible(true);
         }
     }
+
+    public static boolean checkZeroArgConstructor(Class clazz) {
+        try {
+            clazz.getDeclaredConstructor();
+            return true;
+        } catch (NoSuchMethodException e) {
+            return false;
+        }
+    }
+
+    public static boolean isJdk(Class clazz) {
+        return clazz.getName().startsWith("java.") || clazz.getName().startsWith("javax.");
+    }
 }
