+++ dubbo-common/src/test/java/org/apache/dubbo/common/bytecode/ClassGeneratorTest.java
@@ -16,6 +16,8 @@
  */
 package org.apache.dubbo.common.bytecode;
 
+import org.apache.dubbo.common.utils.ReflectUtils;
+
 import org.junit.jupiter.api.Test;
 
 import java.lang.reflect.Field;
@@ -35,7 +37,7 @@ public class ClassGeneratorTest {
         Bean b = new Bean();
         Field fname = null, fs[] = Bean.class.getDeclaredFields();
         for (Field f : fs) {
-            f.setAccessible(true);
+            ReflectUtils.makeAccessible(f);
             if (f.getName().equals("name"))
                 fname = f;
         }
@@ -65,7 +67,7 @@ public class ClassGeneratorTest {
         Bean b = new Bean();
         Field fname = null, fs[] = Bean.class.getDeclaredFields();
         for (Field f : fs) {
-            f.setAccessible(true);
+            ReflectUtils.makeAccessible(f);
             if (f.getName().equals("name"))
                 fname = f;
         }
