+++ dubbo-rpc/dubbo-rpc-api/src/test/java/org/apache/dubbo/rpc/filter/AccessLogFilterTest.java
@@ -19,6 +19,7 @@ package org.apache.dubbo.rpc.filter;
 import org.apache.dubbo.common.URL;
 import org.apache.dubbo.common.utils.DubboAppender;
 import org.apache.dubbo.common.utils.LogUtil;
+import org.apache.dubbo.common.utils.ReflectUtils;
 import org.apache.dubbo.rpc.Filter;
 import org.apache.dubbo.rpc.Invocation;
 import org.apache.dubbo.rpc.Invoker;
@@ -64,7 +65,7 @@ public class AccessLogFilterTest {
         Invocation invocation = new MockInvocation();
 
         Field field = AccessLogFilter.class.getDeclaredField("LOG_ENTRIES");
-        field.setAccessible(true);
+        ReflectUtils.makeAccessible(field);
         assertTrue(((Map) field.get(AccessLogFilter.class)).isEmpty());
 
         accessLogFilter.invoke(invoker, invocation);
