+++ dubbo-rpc/dubbo-rpc-thrift/src/test/java/org/apache/dubbo/rpc/protocol/thrift/ServiceMethodNotFoundTest.java
@@ -17,6 +17,7 @@
 package org.apache.dubbo.rpc.protocol.thrift;
 
 import org.apache.dubbo.common.URL;
+import org.apache.dubbo.common.utils.ReflectUtils;
 import org.apache.dubbo.rpc.gen.dubbo.$__DemoStub;
 import org.apache.dubbo.rpc.gen.dubbo.Demo;
 import org.apache.dubbo.rpc.protocol.thrift.ext.MultiServiceProcessor;
@@ -46,7 +47,7 @@ public class ServiceMethodNotFoundTest extends AbstractTest {
         // for test
         Field field = processor.getClass().getSuperclass().getDeclaredField("processMap");
 
-        field.setAccessible(true);
+        ReflectUtils.makeAccessible(field);
 
         Object obj = field.get(processor);
 
@@ -132,7 +133,7 @@ public class ServiceMethodNotFoundTest extends AbstractTest {
         String arg = "Hello, World!";
 
         invocation.setArguments( new Object[] { arg } );
-        
+
         invocation.setAttachment(Constants.INTERFACE_KEY, DemoImpl.class.getName());
 
         Result result = invoker.invoke( invocation );
